+++ ne_version.h
@@ -1,3 +1,3 @@
 ﻿#pragma once
-#define NE_PUBLISHED L"25.05.2026 14:21"
+#define NE_PUBLISHED L"25.05.2026 14:47"
 #define NE_VERSION   L"2026.05.25.14"